requirements.txt
# Core dependencies
pydantic==2.5.0
aiohttp==3.9.1
python-multipart==0.0.6

# Data processing
pandas>=2.3.0
pyarrow>=14.0.1
numpy>=2.0.0

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
google-cloud-secret-manager==2.16.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Compression
zstandard==0.22.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-benchmark==4.0.0
freezegun==1.2.2
responses==0.24.1

# Development (optional)
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0

# Monitoring (optional)
prometheus-client==0.19.0
psutil==5.9.6

# CLI
click==8.1.7
rich==13.7.0

# Live streaming dependencies
websockets==12.0
asyncio-mqtt==0.16.1
tardis-client
docker
python-dotenv

# BigQuery streaming
google-cloud-bigquery==3.13.0

# HFT features computation
# numpy>=2.0.0  # Already included above
scipy>=1.11.4
scikit-learn>=1.3.2

# Node.js integration (for real-time streaming)
# Note: Requires Node.js runtime and npm install tardis-dev
